refactor(backend): move prover glue + witness solver out of the spine
The provekit-prover crate mixed two concerns: the field-generic WHIR
proving spine (whir_r1cs.rs) and bn254-welded glue (the Prove orchestrator,
the ACIR witness solver, 256-bit bigint/EC arithmetic). The latter cannot
be generic over the proof field, so it does not belong in the spine.

Relocate the Prove trait + impls, solve_witness_vec/CompressedR1CS,
bigint_mod, ec_arith, input_utils, logging, and the witness solver into
provekit-backend-bn254. provekit-prover is reduced to its generic
WhirR1CSProver/MavrosR1CSProver spine and drops its acir/noirc_abi deps.

The dependency edge reverses accordingly: backend now depends on prover
(for the generic proving traits) instead of the reverse. Consumers repoint
Prove/ec_scalar_mul/solve_witness_vec to the backend re-exports; the
witness-generation/parallel feature flags move onto the backend dep.

bn254 output is byte-identical (72/72 e2e green).
