+++ provekit/backend/bn254/Cargo.toml
@@ -8,9 +8,15 @@ license.workspace = true
 homepage.workspace = true
 repository.workspace = true
 
+[features]
+default = ["parallel"]
+witness-generation = ["nargo", "bn254_blackbox_solver", "noir_artifact_cli"]
+parallel = ["provekit-prover/parallel", "provekit-common/parallel"]
+
 [dependencies]
 # Workspace crates
 provekit-common.workspace = true
+provekit-prover.workspace = true
 
 # Noir language
 acir.workspace = true
@@ -19,9 +25,12 @@ noirc_abi.workspace = true
 # Cryptography and proof systems
 ark-bn254.workspace = true
 ark-ff.workspace = true
+ark-std.workspace = true
+whir.workspace = true
 
 # 3rd party
 anyhow.workspace = true
+num-bigint.workspace = true
 postcard.workspace = true
 serde.workspace = true
 serde_json.workspace = true
@@ -29,6 +38,9 @@ tracing.workspace = true
 
 # Target-specific dependencies: only on non-WASM targets
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+bn254_blackbox_solver = { workspace = true, optional = true }
+nargo = { workspace = true, optional = true }
+noir_artifact_cli = { workspace = true, optional = true }
 mavros-vm.workspace = true
 mavros-artifacts.workspace = true
