+++ provekit/backend/bn254/src/witness/digits.rs
