+++ provekit/backend/bn254/src/witness/ram.rs
