+++ provekit/prover/Cargo.toml
@@ -9,18 +9,12 @@ homepage.workspace = true
 repository.workspace = true
 
 [features]
-default = ["witness-generation", "parallel"]
-witness-generation = ["nargo", "bn254_blackbox_solver", "noir_artifact_cli"]
+default = ["parallel"]
 parallel = ["provekit-common/parallel"]
 
 [dependencies]
 # Workspace crates
 provekit-common.workspace = true
-provekit-backend-bn254.workspace = true
-
-# Noir language
-acir.workspace = true
-noirc_abi.workspace = true
 
 # Cryptography and proof systems
 ark-ff.workspace = true
@@ -29,15 +23,10 @@ whir.workspace = true
 
 # 3rd party
 anyhow.workspace = true
-num-bigint.workspace = true
-postcard.workspace = true
 tracing.workspace = true
 
 # Target-specific dependencies: only on non-WASM targets
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-bn254_blackbox_solver = { workspace = true, optional = true }
-nargo = { workspace = true, optional = true }
-noir_artifact_cli = { workspace = true, optional = true }
 mavros-vm.workspace = true
 mavros-artifacts.workspace = true
