+++ tooling/cli/Cargo.toml
@@ -11,9 +11,8 @@ repository.workspace = true
 [dependencies]
 # Workspace crates
 provekit-common.workspace = true
-provekit-backend-bn254.workspace = true
+provekit-backend-bn254 = { workspace = true, features = ["witness-generation", "parallel"] }
 provekit-gnark.workspace = true
-provekit-prover = { workspace = true, features = ["witness-generation", "parallel"] }
 provekit-r1cs-compiler.workspace = true
 provekit-verifier.workspace = true
