+++ tooling/provekit-bench/tests/compiler.rs
@@ -4,9 +4,8 @@ use {
     nargo_cli::cli::compile_cmd::compile_workspace_full,
     nargo_toml::{resolve_workspace_from_toml, PackageSelection},
     noirc_driver::CompileOptions,
-    provekit_backend_bn254::{Prover, Verifier},
+    provekit_backend_bn254::{Prove, Prover, Verifier},
     provekit_common::HashConfig,
-    provekit_prover::Prove,
     provekit_r1cs_compiler::NoirCompiler,
     provekit_verifier::Verify,
     serde::Deserialize,
